requirements.txt
flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
Flask-Migrate
cloudinary
gunicorn